New module with clearly labeled CC generators.
